+++ tests/mir_test_framework/stubbed_server_configuration.cpp
@@ -201,7 +201,7 @@ class StubGraphicPlatform : public mtd::NullPlatform
 class StubRenderer : public mc::Renderer
 {
 public:
-    void begin(int) const override
+    void begin(float) const override
     {
     }
